ALC700: fixed speaker/headphone priority (#494)
Currently the back port (labeled line out, usually connected to speakers)
takes priority over the front port (labeled microphone). That means if
both ports are plugged in, the back port takes priority even though it
makes more sense to have the back port always plugged in to a speaker and
the front port connected to something temporary like headphones. In
practice either port can work, but cable management can get ugly.

This is due to the fact that OSX prioritizes Line Out > HP > Speakers.
The default pin config labels the back port as Line Out but that doesn't
seem to be correct as it has an active amp attached which makes more sense
as a speaker output (line out is supposed to not have any amps). Therefore
this patch re-labels node 0x1B to be a "Speaker" port instead.

OSX also has no concept of an external speaker output, so we have to label
it (in the path file) as an "internal speaker" and it will show up as an
"internal speaker" but there is no way around that.