Add new layout 17 for ALC280 (#591)
This layout is a mix of 13, 15 & 16 such that there will be 2 devices listed in the preferences `Internal Speaker/Line Out` depending on if theres something plugged in the rear green plug, and `Headphones`. You can manually select between these two devices in the preferences or on the sound icon on the top bar.

The main limitations lie with the hardware and with apples decision to have line-out override the headphones. The hardware limitation is that the internal speaker cannot not function if there is something in the line-out plug because they share the same AMP etc. So it only makes sense to have this part auto-switch. And because of the apple decision we really need to manually select that part.